# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StratusCloud Public API
  version: 0.3.0
  description: |
    StratusCloud 1차 MVP 공개 REST API 계약서.
    Week 3에서는 IAM-2 범위(OIDC JWT/API Key/Policy)를 포함한다.
servers:
  - url: https://api.stratuscloud.local
    description: local development gateway
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /v1/projects:
    post:
      tags: [iam]
      summary: 프로젝트를 생성한다.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}:
    get:
      tags: [iam]
      summary: 프로젝트 단건 정보를 조회한다.
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
  /v1/iam/roles:
    post:
      tags: [iam]
      summary: 역할에 정책 목록을 바인딩한다.
      operationId: bindRolePolicies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindRolePoliciesRequest'
      responses:
        '200':
          description: 바인딩 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RolePolicy'
...
  /v1/iam/secrets:
    post:
      tags: [iam]
      summary: Secret을 생성한다.
      operationId: createSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecretRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
    get:
      tags: [iam]
      summary: Secret 목록을 조회한다.
      operationId: listSecrets
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            nullable: true
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Secret'
  /v1/iam/secrets/{secretId}:
    get:
      tags: [iam]
      summary: Secret 메타데이터를 조회한다.
      operationId: getSecret
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
  /v1/iam/secrets/{secretId}/versions:
    get:
      tags: [iam]
      summary: Secret 버전 메타데이터를 조회한다.
      operationId: listSecretVersions
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretVersion'
  /v1/iam/secrets/{secretId}:rotate:
    post:
      tags: [iam]
      summary: Secret을 회전해 새 버전을 생성한다.
      operationId: rotateSecret
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotateSecretRequest'
      responses:
        '200':
          description: 회전 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
  /v1/iam/secrets/{secretId}/versions/{versionId}:
    delete:
      tags: [iam]
      summary: Secret 버전을 회수한다.
      operationId: revokeSecretVersion
      parameters:
        - name: secretId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 회수 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretVersion'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/JWKS 검증을 통과한 JWT 토큰
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 머신 간 호출용 API Key
  schemas:
...
    RoleType:
      type: string
      enum: [OWNER, ADMIN, DEVELOPER, VIEWER]
...
    SecretVersionStatus:
      type: string
      enum: [ACTIVE, REVOKED]
...
    CreateProjectRequest:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 100
    Project:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    BindRolePoliciesRequest:
      type: object
      required: [tenantId, role, policyIds]
      properties:
        tenantId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/RoleType'
        policyIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    RolePolicy:
      type: object
      required: [id, tenantId, role, policyId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/RoleType'
        policyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    CreateSecretRequest:
      type: object
      required: [tenantId, name, value]
      properties:
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          minLength: 2
          maxLength: 100
        value:
          type: string
          minLength: 1
          maxLength: 4096
    RotateSecretRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
          minLength: 1
          maxLength: 4096
    SecretVersion:
      type: object
      required: [id, version, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        status:
          $ref: '#/components/schemas/SecretVersionStatus'
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        value:
          type: string
          nullable: true
          description: 생성/회전 응답에서만 1회 노출되는 값
    Secret:
      type: object
      required: [id, tenantId, name, latestVersion, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        latestVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
        currentVersion:
          allOf:
            - $ref: '#/components/schemas/SecretVersion'
          nullable: true
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          type: object
          additionalProperties: true